+++ L15-structuralPatterns/homework/src/main/java/ru/otus/HomeWork.java
@@ -1,24 +1,41 @@
 package ru.otus;
 
+import ru.otus.handler.ComplexProcessor;
+import ru.otus.listener.homework.HistoryListener;
+import ru.otus.model.Message;
+import ru.otus.model.ObjectForMessage;
+import ru.otus.processor.homework.ProcessorChangeFields;
+import ru.otus.processor.homework.ProcessorTimeException;
+import ru.otus.processor.homework.SystemTimeProvider;
+
+import java.util.ArrayList;
+import java.util.List;
+
 public class HomeWork {
 
-    /*
-     Реализовать to do:
-       1. Добавить поля field11 - field13 (для field13 используйте класс ObjectForMessage)
-       2. Сделать процессор, который поменяет местами значения field11 и field12
-       3. Сделать процессор, который будет выбрасывать исключение в четную секунду (сделайте тест с гарантированным результатом)
-             Секунда должна определяьться во время выполнения.
-             Тест - важная часть задания
-             Обязательно посмотрите пример к паттерну Мементо!
-       4. Сделать Listener для ведения истории (подумайте, как сделать, чтобы сообщения не портились)
-          Уже есть заготовка - класс HistoryListener, надо сделать его реализацию
-          Для него уже есть тест, убедитесь, что тест проходит
-     */
-
     public static void main(String[] args) {
-        /*
-           по аналогии с Demo.class
-           из элеменов "to do" создать new ComplexProcessor и обработать сообщение
-         */
+        var processors = List.of(new ProcessorTimeException(new SystemTimeProvider()),
+                new ProcessorChangeFields());
+
+        var complexProcessor = new ComplexProcessor(processors, ex -> {});
+        var historyListener = new HistoryListener();
+        complexProcessor.addListener(historyListener);
+
+        ObjectForMessage objectForMessage = new ObjectForMessage();
+        objectForMessage.setData(new ArrayList<>());
+
+
+        var message = new Message.Builder(1L)
+                .field1("field1")
+                .field2("field2")
+                .field5("field5")
+                .field9("field9")
+                .field13(objectForMessage)
+                .build();
+
+        var result = complexProcessor.handle(message);
+        System.out.println("result:" + result);
+
+        complexProcessor.removeListener(historyListener);
     }
 }
